refactor(staff): split api.go god object per active-package template (#575 B3)
Pure file moves, no logic changes: api.go (2190 L) → wiring-only
(Resource/NewResource/SetPINValidation/Router, 117 L) + staff_handlers /
substitution_handlers / pin_handlers / schedule_handlers /
time_tracking_handlers + types.go (all DTOs + Bind validators).
Route table unchanged (pinned by the B0 chi.Walk golden). Complexity
ratchet entries re-pathed, scores unchanged.
